+++ CHANGELOG.md
@@ -4,6 +4,16 @@ Notable changes to this project will be documented in this file.
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [4.7.0](https://github.com/dathbe/MMM-MyScoreboard/compare/v4.6.0...v4.7.0) - 2025-04-12
+- **NEW FEATURE**: Australian A-League Women's soccer added
+- BUG FIX: Update README with correct ARI Diamondbacks abbreviation
+- Logic changes to reduce the number of unecessary API calls
+- Under-the-hood changes to make the code more intuitive to me
+- Fix some css errors
+- Add styling to local sports channels (FanDuel and NBC Sports)
+- Update README images
+- Add some broadcast logos
+
 ## [4.6.0](https://github.com/dathbe/MMM-MyScoreboard/compare/v4.5.0...v4.6.0) - 2025-04-09
 - **NEW FEATURE**: New config option to `showLocalBroadcasts`
 - **NEW FEATURE**: New config option to `skipChannels`